Multiple chat rooms at one time now work.
I originally wanted to try to "multiplex" the event stream, but it was
far simpler to simply have the client start new connections from "zero"
(from scratch). I think this is because the client can only listen for
new events, not send information back to the server, except in the very
first request for the event stream itself. Simply having the client look
to see if it's already received a given message (by its ID) makes the
implementation that much easier, at the cost of receiving duplicate
messages on the wire when a new chat room is opened.